pdf_tools_app/requirements.txt
Flask
Flask-Cors
PyPDF2
pdf2image
Pillow
